requirements.txt
streamlit==1.56.0
pandas>=2.2.0
google-generativeai>=0.8.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
fastapi>=0.115.0
uvicorn>=0.30.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
langchain-google-genai>=2.0.0
faiss-cpu>=1.8.0
docx2txt>=0.8
pypdf>=4.0.0
